bmi-langgraph_app/requirements.txt
langgraph
langchain-google-genai
streamlit
python-dotenv
